requirements.txt
aws-cdk.assets>=1.32.2
aws-cdk.aws-apigateway>=1.32.2
aws-cdk.aws-applicationautoscaling>=1.32.2
aws-cdk.aws-autoscaling>=1.32.2
aws-cdk.aws-autoscaling-common>=1.32.2
aws-cdk.aws-autoscaling-hooktargets>=1.32.2
aws-cdk.aws-batch>=1.32.2
aws-cdk.aws-certificatemanager>=1.32.2
aws-cdk.aws-cloudformation>=1.32.2
aws-cdk.aws-cloudfront>=1.32.2
aws-cdk.aws-cloudwatch>=1.32.2
aws-cdk.aws-codebuild>=1.32.2
aws-cdk.aws-codecommit>=1.32.2
aws-cdk.aws-codepipeline>=1.32.2
aws-cdk.aws-cognito>=1.32.2
aws-cdk.aws-dynamodb>=1.32.2
aws-cdk.aws-ec2>=1.32.2
aws-cdk.aws-ecr>=1.32.2
aws-cdk.aws-ecr-assets>=1.32.2
aws-cdk.aws-ecs>=1.32.2
aws-cdk.aws-elasticloadbalancing>=1.32.2
aws-cdk.aws-elasticloadbalancingv2>=1.32.2
aws-cdk.aws-events>=1.32.2
aws-cdk.aws-events-targets>=1.32.2
aws-cdk.aws-glue>=1.32.2
aws-cdk.aws-iam>=1.32.2
aws-cdk.aws-kinesis>=1.32.2
aws-cdk.aws-kms>=1.32.2
aws-cdk.aws-lambda>=1.32.2
aws-cdk.aws-lambda-event-sources>=1.32.2
aws-cdk.aws-logs>=1.32.2
aws-cdk.aws-route53>=1.32.2
aws-cdk.aws-route53-targets>=1.32.2
aws-cdk.aws-s3>=1.32.2
aws-cdk.aws-s3-assets>=1.32.2
aws-cdk.aws-s3-notifications>=1.32.2
aws-cdk.aws-sagemaker>=1.32.2
aws-cdk.aws-sam>=1.32.2
aws-cdk.aws-secretsmanager>=1.32.2
aws-cdk.aws-servicediscovery>=1.32.2
aws-cdk.aws-sns>=1.32.2
aws-cdk.aws-sns-subscriptions>=1.32.2
aws-cdk.aws-sqs>=1.32.2
aws-cdk.aws-ssm>=1.32.2
aws-cdk.aws-stepfunctions>=1.32.2
aws-cdk.aws-stepfunctions-tasks>=1.32.2
aws-cdk.core>=1.32.2
aws-cdk.custom-resources>=1.32.2
aws-cdk.cx-api>=1.32.2
aws-cdk.region-info>=1.32.2
boto3>=1.12.39
botocore>=1.15.39
flake8>=3.7.9
pytest>=5.4.1

